Folder creation and do not overwrite files

+++ StateMachineGeneretorV2/Base/BaseGenerator.cs
@@ -9,8 +9,15 @@ namespace StateMachineGeneretorV2.Base
     {
         protected static void CreateFile(string pathString, string fileContent)
         {
-            File.WriteAllText(pathString, fileContent);
-            Console.WriteLine("File {0} written.", pathString.Replace("\\", "/"));
+            if (!File.Exists(pathString))
+            {
+                File.WriteAllText(pathString, fileContent);
+                Console.WriteLine("File {0} written.", pathString.Replace("\\", "/"));
+            }
+            else 
+            {
+                Console.WriteLine("File {0} already exist.", pathString.Replace("\\", "/"));
+            }
         }
     }
 }
